Sync the latest version
First attempt to use myweather2.com to generate snow forecasts.  The
current mechanism is not going to work as it involves cross site
scripting.  The only feasible solution to this is to implement a server
side proxy, which means changing hosts.  Not sure I can be bothered to
do that, but we'll see how bored I get.

SnowReport still uses an iframe to render the snow reports.
WebServiceSnowReport in theory will render out data received from the
web service.